Add dom manipulation for await of fetch
