+++ src/index.html
@@ -6,6 +6,7 @@
 </head>
 <body>
   <h1>Leaderboard</h1>
+  <h2 class="processing hidden" id="processing">Processing Request...</h2>
   <div class="content">
     <div class="left-content">
       <div class="table-header">
@@ -26,7 +27,7 @@
         <input class="form-field" name="score" id="score" placeholder="Your score" required>
         <br>
         <div class="submit-btn">
-          <button type="submit" class="add-score-btn">Submit</button>
+          <button type="button" class="add-score-btn" id="submit-score">Submit</button>
         </div> 
       </form>
     </div>
